tweak saidahs project
