+++ .gitignore
@@ -1,2 +1,3 @@
 ibm-credentials.env
 node_modules
+.env
