--- bin/index.js
@@ -1,9 +1,0 @@
-#!/usr/bin/env node
-
-//console.log( "Hello World!" );
-
-//Grab provided args.
-const [,, ...args] = process.argv
-
-//Print hello world provided args.
-console.log (`Hello World ${args}`)
