+++ bin/runquery.js
@@ -2,4 +2,4 @@
 //console.log('testing')
 var shell = require("shelljs");
 
-shell.exec("node AddDocs.js");
+shell.exec("node ./src/Query.js");
