+++ package.json
@@ -7,8 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "bin": {
-    "hello": "./bin/index.js",
-    "AddDoc": "./bin/runquery.js"
+    "getTopics": "./bin/runquery.js"
   },
   "repository": {
     "type": "git",
